Adds nix support
We explicitly do not use the agda and stdlib from nixpkgs. PLFA
generally expects exact versions, but we would be at the mercy of
whatever nixpkgs provides.

Instead, we get agda from hackage and use the stdlib from the
submodule. This should be more robust wrt future updates.

We also add CI for verifying this workflow (and delete other CI).